[upstream] Sensor hits (box2d #945)
[upstream] https://github.com/erincatto/box2d/pull/945

Sensor hits for shapes that move fast over sensors.
`b2SensorData` that has the visitor shape id and the captured visitor
body transform.

Bug fix: sensors were not using the custom filter.
Time of impact now returns the point and normal.
`b2PreSolveFcn` now returns a contact point and normal instead of the
manifold. This change was made to support more efficient continuous
collision and for consistency. Pre-solve is called in continuous and
discrete collision, but continuous collision no longer uses manifolds.
Contact push speed is now independent of contact hertz and damping
ratio.
Bumped version to 3.2.0 (work in progress).

Renamed: `b2Shape_GetMassData` to `b2Shape_ComputeMassData`
Removed: `b2Contact_GetManifold`  and `b2Contact_GetShapeIds`
Added: `b2Contact_GetData`
Changed function order on some collision functions to assist editor
auto-complete
`id.h` is now stand-alone
Renamed: `b2WorldDef::maxContactPushSpeed` to `contactSpeed`
Added: `b2BodyDef::enableSensorHits`
Added: `b2JointDef::constraintHertz` and `b2JointDef::constraintDampingRatio`
Renamed: `b2JointDef::drawSize` to `b2JointDef::drawScale`
Removed: `b2ContactBeginTouchEvent::manifold`
Added: `b2ContactData::contactId`
Modified: `b2PreSolveFcn` receives a point and normal instead of a manifold

5st

+++ src/Box2D.NET.Samples/Camera.cs
@@ -12,8 +12,8 @@ public class Camera
 {
     public B2Vec2 m_center;
     public float m_zoom;
-    public int m_width;
-    public int m_height;
+    public float m_width;
+    public float m_height;
 
     public Camera()
     {
